requirements.txt
PyYAML>=6.0
jieba>=0.42
rank_bm25>=0.2.2
pytest>=7.0
pytest-cov>=4.0
